- add order repo

+++ src/main/java/org/example/controller/OrderController.java
@@ -1,4 +1,6 @@
 package org.example.controller;
+import org.example.service.OrderService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
@@ -28,7 +30,7 @@ public class OrderController {
     }
 
     @PostMapping(name = "/checkout")
-    public void checkOut(Long userId){
+    public void checkOut(int userId){
         orderService.checkOut(userId);
     }
 }
