+++ src/main/java/org/example/repository/OrderRepo.java
@@ -9,9 +9,9 @@ import org.example.enums.OrderStatus;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.hibernate.query.Query;
+
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.transaction.annotation.Transactional;
+
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -55,7 +55,7 @@ public class OrderRepo {
         return false;
     }
 
-    public void checkOut(Long userId) {
+    public void checkOut(int userId) {
         //***summary***
         //create an order
         //loop
